+++ ajax/ajaxdecktype.php
@@ -1,8 +1,8 @@
 <?php
 
 /*
-Version:     1.12
-Date:        10/01/26
+Version:     1.13
+Date:        11/01/26
 Name:        ajaxdecktype.php
 Purpose:     AJAX deck type updates for deck detail.
 Notes:       The page does not run standard secpagesetup as it breaks the ajax login catch.
@@ -76,7 +76,7 @@ if (!in_array($updateType, $validtypes, true)) :
     returnResponse($response);
 endif;
 
-$sessionManager = new SessionManager($db, $adminip, $_SESSION, $fxAPI, $fxLocal, $logfile);
+$sessionManager = new SessionManager($db, $_SESSION, $appConfig);
 $userArray = $sessionManager->getUserInfo();
 $user = $userArray['usernumber'];
 $userEmail = $_SESSION['useremail'];